af-packet: use per CPU hash in bypass
eBPF has a data type which is a per CPU array. By adding one element
to the array it is in fact added to all per CPU arrays in the kernel.
This allows to have a lockless structure in the kernel even when doing
counter update.

In userspace, we need to update the flow bypass code to fetch all
elements of the per CPU arrays.

+++ src/source-af-packet.c
@@ -246,7 +246,7 @@ typedef struct AFPThreadVars_
     /* handle state */
     uint8_t afp_state;
     uint8_t copy_mode;
-    uint8_t flags;
+    int flags;
 
     /* IPS peer */
     AFPPeer *mpeer;
